+++ Bot/Cogs/kumiko/economy/auction_house.py
@@ -1,28 +1,20 @@
 import asyncio
 import os
-import re
-import uuid
-from datetime import datetime
 
-import aiormq
 import discord
 import uvloop
-import yaml
 from dateutil import parser
 from discord.commands import Option, SlashCommandGroup
 from discord.ext import commands, pages
 from dotenv import load_dotenv
 from economy_utils import KumikoAuctionHouseUtils, KumikoEcoUserUtils
-from kumiko_ui_components import AHPurgeAllView
+from kumiko_ui_components import AHCreateItemModal, AHDeleteItemModal, AHPurgeAllView
 from rin_exceptions import ItemNotFound, NoItemsError
 
 load_dotenv()
 
-RABBITMQ_USER = os.getenv("RabbitMQ_Username")
-RABBITMQ_PASSWORD = os.getenv("RabbitMQ_Password")
-RABBITMQ_SERVER_IP = os.getenv("RabbitMQ_Server_IP")
 REDIS_SERVER_IP = os.getenv("Redis_Server_IP")
-REDIS_SERVER_PORT = os.getenv("Redis_Port_Dev")
+REDIS_SERVER_PORT = os.getenv("Redis_Port")
 POSTGRES_PASSWORD = os.getenv("Postgres_Password")
 POSTGRES_SERVER_IP = os.getenv("Postgres_Server_IP")
 POSTGRES_DB = os.getenv("Postgres_Kumiko_Database")
@@ -35,63 +27,6 @@ auctionHouseUtils = KumikoAuctionHouseUtils()
 userUtils = KumikoEcoUserUtils()
 
 
-class View(discord.ui.View):
-    async def on_timeout(self):
-        for child in self.children:
-            child.disabled = True
-
-    @discord.ui.button(
-        label="Yes",
-        row=0,
-        style=discord.ButtonStyle.primary,
-        emoji=discord.PartialEmoji.from_str("<:check:314349398811475968>"),
-    )
-    async def button_callback(self, button, interaction):
-        try:
-            getUserInfo = await userUtils.selectUserRank(
-                interaction.user.id, CONNECTION_URI
-            )
-            if len(getUserInfo) == 0:
-                raise NoItemsError
-            else:
-                for items in getUserInfo:
-                    if items < 25:
-                        await interaction.response.send_message(
-                            f"Sorry, but your current rank is {items}. You need at the very least rank 25 or higher to use this command."
-                        )
-                    else:
-                        itemUUIDAuth = await auctionHouseUtils.obtainItemUUIDAuth(
-                            user_id=interaction.user.id, uri=AH_CONNECTION_URI
-                        )
-                        try:
-                            if len(itemUUIDAuth) == 0:
-                                raise ItemNotFound
-                            else:
-                                await auctionHouseUtils.purgeUserAHItems(
-                                    user_id=interaction.user.id, uri=AH_CONNECTION_URI
-                                )
-                                await interaction.response.send_message(
-                                    "Confirmed. All Auction House Listings have now been completely purged from your account. This is permanent and irreversible."
-                                )
-                        except ItemNotFound:
-                            await interaction.response.send_message(
-                                "It seems like you don't have any to delete from at all..."
-                            )
-        except NoItemsError:
-            await interaction.response.send_message(
-                "It seems like you don't even have an account to begin with. Go ahead and create one first."
-            )
-
-    @discord.ui.button(
-        label="No",
-        row=0,
-        style=discord.ButtonStyle.primary,
-        emoji=discord.PartialEmoji.from_str("<:xmark:314349398824058880>"),
-    )
-    async def second_button_callback(self, button, interaction):
-        await interaction.response.send_message("Well glad you choose not to...")
-
-
 class AuctionHouse(commands.Cog):
     """A place reserved for the elites for auctioning high value items"""
 
@@ -101,10 +36,9 @@ class AuctionHouse(commands.Cog):
     auctionHouse = SlashCommandGroup(
         "eco-auction",
         "Auction off your highest items here",
-        guild_ids=[970159505390325842],
     )
     auctionHouseDelete = auctionHouse.create_subgroup(
-        "delete", "commands to delete stuff", guild_ids=[970159505390325842]
+        "delete", "commands to delete stuff"
     )
 
     @auctionHouse.command(name="select")
@@ -113,24 +47,28 @@ class AuctionHouse(commands.Cog):
     ):
         """Selects an item on the Auction House before to bid for"""
         try:
-            getUserInfo = await userUtils.selectUserRank(ctx.author.id, CONNECTION_URI)
-            if len(getUserInfo) == 0:
+            getUserInfo = await userUtils.getFirstUser(
+                user_id=ctx.author.id, uri=CONNECTION_URI
+            )
+            if getUserInfo is None:
                 raise ItemNotFound
             else:
-                for items in getUserInfo:
-                    if items < 25:
+                if dict(getUserInfo)["rank"] < 25:
+                    await ctx.respond(
+                        f"Sorry, but your current rank is {dict(getUserInfo)['rank']}. You need at the very least rank 25 in order to use this command."
+                    )
+                else:
+                    selectMainItem = await auctionHouseUtils.selectFirstItem(
+                        name=name, uri=AH_CONNECTION_URI
+                    )
+                    if selectMainItem is None:
                         await ctx.respond(
-                            f"Sorry, but your current rank is {items}. You need at the very least rank 25 in order to use this command."
+                            "It seems like the item requested could not be found. Please try again later"
                         )
                     else:
-                        selectMainItem = await auctionHouseUtils.selectAHItemUUID(
-                            name=name, uri=AH_CONNECTION_URI
-                        )
-                        for item in selectMainItem:
-                            ahItemUUID = item
                         await auctionHouseUtils.setItemKey(
-                            str(ctx.user.id),
-                            ahItemUUID,
+                            key=str(ctx.user.id),
+                            value=dict(selectMainItem)["uuid"],
                             db=1,
                             ttl=21660,
                             redis_server_ip=REDIS_SERVER_IP,
@@ -148,60 +86,53 @@ class AuctionHouse(commands.Cog):
     async def bindForPrice(self, ctx, *, price: Option(int, "The price to bid")):
         """Bid for a higher price for an item"""
         try:
-            try:
-                getUserInfo = await userUtils.selectUserRank(
-                    ctx.author.id, CONNECTION_URI
-                )
-                if len(getUserInfo) == 0:
-                    raise ItemNotFound
+            getUserInfo = await userUtils.getFirstUser(
+                user_id=ctx.author.id, uri=CONNECTION_URI
+            )
+            if getUserInfo is None:
+                raise ItemNotFound
+            else:
+                if dict(getUserInfo)["rank"] < 25:
+                    await ctx.respond(
+                        f"Sorry, but your current rank is {dict(getUserInfo)['rank']}. You need at the very least rank 25 in order to use this command."
+                    )
                 else:
-                    for items in getUserInfo:
-                        if items < 25:
-                            await ctx.respond(
-                                f"Sorry, but your current rank is {mainItems['rank']}. You need at the very least rank 25 in order to use this command."
+                    selectedItem = await auctionHouseUtils.getItemKey(
+                        key=str(ctx.user.id),
+                        db=1,
+                        redis_server_ip=REDIS_SERVER_IP,
+                        redis_port=REDIS_SERVER_PORT,
+                    )
+                    if selectedItem is None:
+                        await ctx.respond(
+                            "You have not selected an item yet. Please run the command to select an item first."
+                        )
+                    else:
+                        getSelectedItem = (
+                            await auctionHouseUtils.selectFirstItemViaUUID(
+                                uuid=selectedItem, uri=AH_CONNECTION_URI
                             )
-                        else:
-                            selectedItem = await auctionHouseUtils.getItemKey(
-                                str(ctx.user.id),
-                                db=1,
-                                redis_server_ip=REDIS_SERVER_IP,
-                                redis_port=REDIS_SERVER_PORT,
-                            )
-                            if selectedItem is None:
-                                await ctx.respond(
-                                    "You have not selected an item yet. Please run the command to select an item first."
-                                )
+                        )
+                        try:
+                            if price < int(dict(getSelectedItem)["price"]):
+                                raise ValueError
                             else:
-                                selectItemPriceRes = (
-                                    await auctionHouseUtils.selectAHItemPrice(
-                                        uuid=selectedItem, uri=AH_CONNECTION_URI
-                                    )
+                                await auctionHouseUtils.setItemKey(
+                                    key=f"{ctx.user.id}-{selectedItem}-bid",
+                                    value=price,
+                                    db=1,
+                                    ttl=3600,
+                                    redis_server_ip=REDIS_SERVER_IP,
+                                    redis_port=REDIS_SERVER_PORT,
                                 )
-                                for mainItems in selectItemPriceRes:
-                                    selectItemPrice = mainItems
-                                if price < int(selectItemPrice):
-                                    raise ValueError
-                                else:
-                                    connection = await aiormq.connect(
-                                        f"amqp://{RABBITMQ_USER}:{RABBITMQ_PASSWORD}@{RABBITMQ_SERVER_IP}/"
-                                    )
-                                    channel = await connection.channel()
-                                    await channel.exchange_declare(
-                                        exchange="auction_house", exchange_type="fanout"
-                                    )
-                                    body = bytes(f"{selectedItem}:{price}", "utf-8")
-                                    await channel.basic_publish(
-                                        body, exchange="auction_house"
-                                    )
-                                    await connection.close()
-                                    await ctx.respond(f"bid placed ({price})")
-            except ItemNotFound:
-                await ctx.respond(
-                    "It seems like you don't even have an account to begin with. Go ahead and create one first."
-                )
-        except ValueError:  # THIS NEEDS TO BE CHANGED OUT WITH A CUSTOM EXCEPTION LATER
+                                await ctx.respond(f"Bid set for {price}")
+                        except ValueError:
+                            await ctx.respond(
+                                "It seems like the price given is lower than the one set. Please try again."
+                            )
+        except ItemNotFound:
             await ctx.respond(
-                "It seems like the price given is lower than the one set. Please try again."
+                "It seems like you don't even have an account to begin with. Go ahead and create one first."
             )
 
     asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
@@ -210,85 +141,15 @@ class AuctionHouse(commands.Cog):
     async def addItemToAuctionHouse(
         self,
         ctx,
-        *,
-        name: Option(str, "The item that you are putting up for bidding"),
-        description: Option(str, "The description of the item"),
-        price: Option(int, "The initial price of the item"),
     ):
         """Adds an item to the Auction House"""
-        dateAdded = datetime.now().isoformat()
-        auctionHouseItemUUID = str(uuid.uuid4())
-        userID = ctx.user.id
-        filterFile = os.path.join(
-            os.path.dirname(__file__), "config", "marketplace_filters.yml"
+        mainModal = AHCreateItemModal(
+            uri=AH_CONNECTION_URI,
+            redis_host=REDIS_SERVER_IP,
+            redis_port=REDIS_SERVER_PORT,
+            title="Create an item",
         )
-        try:
-            try:
-                getUserInfo = await userUtils.obtainUserData(
-                    ctx.author.id, CONNECTION_URI
-                )
-                if len(getUserInfo) == 0:
-                    raise ItemNotFound
-                else:
-                    for items in getUserInfo:
-                        mainItems = dict(items)
-                    if mainItems["rank"] < 25:
-                        await ctx.respond(
-                            f"Sorry, but your current rank is {mainItems['rank']}. You need at the very least rank 25 in order to use this command."
-                        )
-                    elif mainItems["rank"] > 25 and mainItems["lavender_petals"] < 1500:
-                        await ctx.respond(
-                            "It seems like you don't have enough money to list. You need at the very least 1500 Petals to list"
-                        )
-                    elif (
-                        mainItems["rank"] > 25 and mainItems["lavender_petals"] >= 1500
-                    ):
-                        with open(filterFile, "r") as stream:
-                            try:
-                                filterList = yaml.safe_load(stream)
-                                mainFilterList = filterList["filters"]
-                                mainFilter = re.compile(
-                                    "|".join([str(item) for item in mainFilterList]),
-                                    re.IGNORECASE,
-                                )
-                                filteredItemName = re.sub(mainFilter, "item", str(name))
-                                filteredItemDescription = re.sub(
-                                    mainFilter, "item", str(description)
-                                )
-                                await auctionHouseUtils.addAuctionHouseItem(
-                                    auctionHouseItemUUID,
-                                    userID,
-                                    filteredItemName,
-                                    filteredItemDescription,
-                                    dateAdded,
-                                    price,
-                                    False,
-                                    AH_CONNECTION_URI,
-                                )
-                                await auctionHouseUtils.setItemKey(
-                                    auctionHouseItemUUID,
-                                    price,
-                                    db=0,
-                                    ttl=86400,
-                                    redis_server_ip=REDIS_SERVER_IP,
-                                    redis_port=REDIS_SERVER_PORT,
-                                )
-                                await ctx.respond("Item added to the auction house")
-                            except yaml.YAMLError:
-                                await ctx.respond(
-                                    "Oops, something went wrong with the yaml file! Please try again."
-                                )
-
-            except ItemNotFound:
-                await ctx.respond(
-                    "It seems like you don't even have an account to begin with. Go ahead and create one first."
-                )
-        except Exception:
-            await ctx.respond(
-                embed=discord.Embed(
-                    description="Oops, something went wrong! Please try again."
-                )
-            )
+        await ctx.send_modal(mainModal)
 
     asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
 
@@ -296,52 +157,51 @@ class AuctionHouse(commands.Cog):
     async def viewAHItems(self, ctx):
         """Views all of the items on the AH for bidding"""
         try:
-            getUserInfo = await userUtils.selectUserRank(ctx.author.id, CONNECTION_URI)
-            if len(getUserInfo) == 0:
+            getUserInfo = await userUtils.getFirstUser(
+                user_id=ctx.author.id, uri=CONNECTION_URI
+            )
+            if getUserInfo is None:
                 raise NoItemsError
             else:
-                for items in getUserInfo:
-                    if items < 25:
-                        await ctx.respond(
-                            f"Sorry, but your current rank is {items}. You need at the very least rank 25 or higher to use this command."
-                        )
-                    else:
-                        mainRes = await auctionHouseUtils.selectItemNotPassed(
-                            False, AH_CONNECTION_URI
-                        )
-                        try:
-                            if len(mainRes) == 0:
-                                raise ItemNotFound
-                            else:
-                                mainPages = pages.Paginator(
-                                    pages=[
-                                        discord.Embed(
-                                            title=dict(items)["name"],
-                                            description=dict(items)["description"],
-                                        )
-                                        .add_field(
-                                            name="Date Added",
-                                            value=parser.isoparse(
-                                                dict(items)["date_added"]
-                                            ).strftime("%Y-%m-%d %H:%M:%S"),
-                                            inline=True,
-                                        )
-                                        .add_field(
-                                            name="Price",
-                                            value=dict(items)["price"],
-                                            inline=True,
-                                        )
-                                        for items in mainRes
-                                    ],
-                                    loop_pages=True,
-                                )
-                                await mainPages.respond(
-                                    ctx.interaction, ephemeral=False
-                                )
-                        except ItemNotFound:
-                            embedError = discord.Embed()
-                            embedError.description = "There are no items to be seen on the Auction House. Come back later for more!"
-                            await ctx.respond(embed=embedError)
+                if dict(getUserInfo)["rank"] < 25:
+                    await ctx.respond(
+                        f"Sorry, but your current rank is {dict(getUserInfo)['rank']}. You need at the very least rank 25 or higher to use this command."
+                    )
+                else:
+                    mainRes = await auctionHouseUtils.selectItemNotPassed(
+                        passed=False, uri=AH_CONNECTION_URI
+                    )
+                    try:
+                        if len(mainRes) == 0:
+                            raise ItemNotFound
+                        else:
+                            mainPages = pages.Paginator(
+                                pages=[
+                                    discord.Embed(
+                                        title=dict(items)["name"],
+                                        description=dict(items)["description"],
+                                    )
+                                    .add_field(
+                                        name="Date Added",
+                                        value=parser.isoparse(
+                                            dict(items)["date_added"]
+                                        ).strftime("%Y-%m-%d %H:%M:%S"),
+                                        inline=True,
+                                    )
+                                    .add_field(
+                                        name="Price",
+                                        value=dict(items)["price"],
+                                        inline=True,
+                                    )
+                                    for items in mainRes
+                                ],
+                                loop_pages=True,
+                            )
+                            await mainPages.respond(ctx.interaction, ephemeral=False)
+                    except ItemNotFound:
+                        embedError = discord.Embed()
+                        embedError.description = "There are no items to be seen on the Auction House. Come back later for more!"
+                        await ctx.respond(embed=embedError)
         except NoItemsError:
             await ctx.respond(
                 "It seems like you don't even have an account to begin with. Go ahead and create one first."
@@ -350,43 +210,10 @@ class AuctionHouse(commands.Cog):
     asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
 
     @auctionHouseDelete.command(name="one")
-    async def userDelOne(
-        self, ctx, *, name: Option(str, "The name of the item to delete")
-    ):
+    async def userDelOne(self, ctx):
         """Deletes one user from the user for the Auction House"""
-        try:
-            getUserInfo = await userUtils.selectUserRank(ctx.author.id, CONNECTION_URI)
-            if len(getUserInfo) == 0:
-                raise NoItemsError
-            else:
-                for items in getUserInfo:
-                    if items < 25:
-                        await ctx.respond(
-                            f"Sorry, but your current rank is {items}. You need at the very least rank 25 or higher to use this command."
-                        )
-                    else:
-                        userSelRes = await auctionHouseUtils.selectUserItemViaName(
-                            ctx.author.id, name, AH_CONNECTION_URI
-                        )
-                        try:
-                            if len(userSelRes) == 0:
-                                raise ItemNotFound
-                            else:
-                                for items in userSelRes:
-                                    mainItems = dict(items)
-                                    itemUUID = mainItems["uuid"]
-                                    await auctionHouseUtils.deleteUserAHItem(
-                                        ctx.author.id, itemUUID, AH_CONNECTION_URI
-                                    )
-                                await ctx.respond("Item deleted")
-                        except ItemNotFound:
-                            embedError = discord.Embed()
-                            embedError.description = f"There are no items within the Auction House that has the name of {name}. Please try again."
-                            await ctx.respond(embed=embedError)
-        except NoItemsError:
-            await ctx.respond(
-                "It seems like you don't even have an account to begin with. Go ahead and create one first."
-            )
+        mainModal = AHDeleteItemModal(uri=AH_CONNECTION_URI, title="Delete an item")
+        await ctx.send_modal(mainModal)
 
     asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
 
@@ -395,7 +222,9 @@ class AuctionHouse(commands.Cog):
         """Purges all of the user's listed Auction House items. This cannot be recovered from"""
         embed = discord.Embed()
         embed.description = "Are you sure you want to purge all of your Auction House listings? This is permanent and cannot be undone."
-        await ctx.respond(embed=embed, view=AHPurgeAllView(uri=AH_CONNECTION_URI))
+        await ctx.respond(
+            embed=embed, view=AHPurgeAllView(uri=AH_CONNECTION_URI), ephemeral=True
+        )
 
     asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())